update: db用queryで使用しているinterfaceを別ファイル化
