+++ src/db/query/inteface.ts
@@ -1,0 +1,12 @@
+export interface QueryFindUser {
+  realName: string;
+  displayName: string;
+  slackID: string;
+}
+
+export interface QueryFindMessage {
+  ts: string;
+  content: string;
+  userId: string;
+  channelId: string;
+}
