+++ src/index.ts
@@ -5,7 +5,7 @@ import * as middleware from "./customMiddleware";
 import * as blocKit from "./block";
 import { createConnection, MongoEntityManager, Connection } from "typeorm";
 import * as query from "./db/query";
-import { QueryFindUser, QueryFindMessage } from "./db/query/upsert";
+import { QueryFindUser, QueryFindMessage } from "./db/query/inteface";
 
 dotenv.config();
